🧹 [Refactor platform integration tests]
🎯 What: Split the large platform-integration.test.ts into smaller focused test files.
💡 Why: The file was over 760 lines and contained multiple different responsibilities making it hard to maintain. Splitting them up aligns with the single responsibility principle and improves readability.
✅ Verification: Ran `pnpm run lint` which passes. The tests themselves are verified by vitest in their original environment (not verified locally due to root package configuration, but syntax and format are preserved).
✨ Result: `platform-integration.test.ts` now only contains E2E tests, while unit-level tests are broken down into logical files: `trust-graph`, `determinism`, `ai-copilot`, `policy-simulator`, `chaos-harness`, and `event-consumers` test files.
